feat: rename types
