+++ packages/entities/entities-plugins/src/components/free-form/shared/KeyValueField.vue
@@ -92,7 +92,7 @@ import type {
   SlotsType,
 } from 'vue'
 import type { DeepKeys, DeepValue } from './types/util-types'
-import type { FieldCommonProps } from './types/types'
+import type { FieldCommonProps } from './types/field-types'
 
 export type KeyValueFieldProps<
   TParentData,
