+++ packages/entities/entities-plugins/src/components/free-form/shared/RedisSelector.vue
@@ -139,7 +139,7 @@ import { FORMS_CONFIG, REDIS_PARTIAL_FETCHER_KEY } from '@kong-ui-public/forms'
 import { AddIcon } from '@kong/icons'
 import type { SelectItem } from '@kong/kongponents/dist/types'
 import { useAxios, useDebouncedFilter, useErrors, type KongManagerBaseFormConfig, type KonnectBaseFormConfig } from '@kong-ui-public/entities-shared'
-import type { RedisConfig, RedisPartialType, Redis } from './types/types'
+import type { RedisConfig, RedisPartialType, Redis } from './types/field-types'
 import { partialEndpoints, fieldsOrder, REDIS_PARTIAL_INFO } from './const'
 import { getRedisType, getPartialTypeDisplay } from './utils'
 import { useField, useFormData } from './composables'
